fix: Two DbContexts merged into a single context
